+++ lib/skinbaron_api_client/logger.rb
@@ -49,10 +49,14 @@ module SkinbaronApiClient
         @request_logger.info("\n\n#{"=" * 80}")
         @request_logger.info(JSON.pretty_generate(log_data))
       when "RESPONSE"
+        log_data[:body] = JSON.parse(log_data[:body])
         @request_logger.info("\n#{"-" * 80}")
         @request_logger.info(JSON.pretty_generate(log_data))
         @request_logger.info("#{"=" * 80}\n")
       end
+    rescue JSON::ParserError => e
+      @logger.error("LOGGER: failed to parse search response", error: e.message)
+      @logger.error("LOGGER: Response: #{JSON.pretty_generate(log_data)}")
     end
 
     LEVELS.each do |level|